+++ pkg/codegen/internal/test/testdata/simple-enum-schema/nodejs/index.ts
@@ -21,10 +21,10 @@ export {
 
 import { Provider } from "./provider";
 
-pulumi.runtime.registerResourcePackage("plant-provider", {
+pulumi.runtime.registerResourcePackage("plant", {
     version: utilities.getVersion(),
     constructProvider: (name: string, type: string, urn: string): pulumi.ProviderResource => {
-        if (type !== "pulumi:providers:plant-provider") {
+        if (type !== "pulumi:providers:plant") {
             throw new Error(`unknown provider type ${type}`);
         }
         return new Provider(name, <any>undefined, { urn });